+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - fetchActiveModuleActionsFromHttpResponse
 - Token Bearer to doHttpRequest
 - Project Manager Button
+- Project Manager Button: Developer, Pilot, Final and BK-Admin Views